Setup CI caching (#309)
This PR speeds up CI by caching dependencies and using ccache (which is
persisted across runs in addition to significantly helping within a run)

+++ ci/setup-prerequisites-unittest.sh
@@ -1,39 +1,44 @@
 #!/bin/bash
 
-mkdir zstd
-cd zstd
-git init
-git remote add origin https://github.com/facebook/zstd.git
-git fetch --depth 1 origin 794ea1b0afca0f020f4e57b6732332231fb23c70 # 1.5.6
-git checkout FETCH_HEAD
-make -j
-sudo make install
+if [ ! -f zstd/lib/libzstd.a ]; then
+    mkdir zstd
+    cd zstd
+    git init
+    git remote add origin https://github.com/facebook/zstd.git
+    git fetch --depth 1 origin 794ea1b0afca0f020f4e57b6732332231fb23c70 # 1.5.6
+    git checkout FETCH_HEAD
+    make -j
+    cd ..
+fi
+cd zstd && sudo make install && cd ..
 
-cd ..
+if [ ! -f libdwarf/build/src/lib/libdwarf/libdwarf.a ]; then
+    mkdir libdwarf
+    cd libdwarf
+    git init
+    git remote add origin https://github.com/jeremy-rifkin/libdwarf-lite.git
+    git fetch --depth 1 origin 5dfb2cd2aacf2bf473e5bfea79e41289f88b3a5f # 2.1.0
+    git checkout FETCH_HEAD
+    mkdir build
+    cd build
+    cmake .. -GNinja -DPIC_ALWAYS=TRUE -DBUILD_DWARFDUMP=FALSE
+    cd ../..
+fi
+cd libdwarf/build && sudo ninja install && cd ../..
 
-mkdir libdwarf
-cd libdwarf
-git init
-git remote add origin https://github.com/jeremy-rifkin/libdwarf-lite.git
-git fetch --depth 1 origin 5dfb2cd2aacf2bf473e5bfea79e41289f88b3a5f # 2.1.0
-git checkout FETCH_HEAD
-mkdir build
-cd build
-cmake .. -GNinja -DPIC_ALWAYS=TRUE -DBUILD_DWARFDUMP=FALSE
-sudo ninja install
-
-cd ../..
-
-mkdir googletest
-cd googletest
-git init
-git remote add origin https://github.com/google/googletest.git
-git fetch --depth 1 origin f8d7d77c06936315286eb55f8de22cd23c188571
-git checkout FETCH_HEAD
-mkdir build
-cd build
-cmake .. -GNinja -DCMAKE_INSTALL_PREFIX=/tmp/gtest_install
-sudo ninja install
-rm -rf *
-cmake .. -GNinja -DCMAKE_CXX_COMPILER=clang++-18 -DCMAKE_CXX_FLAGS=-stdlib=libc++ -DCMAKE_INSTALL_PREFIX=/tmp/gtest_install_libcxx
-sudo ninja install
+if [ ! -d /tmp/gtest_install ] || [ ! -d /tmp/gtest_install_libcxx ]; then
+    mkdir googletest
+    cd googletest
+    git init
+    git remote add origin https://github.com/google/googletest.git
+    git fetch --depth 1 origin f8d7d77c06936315286eb55f8de22cd23c188571
+    git checkout FETCH_HEAD
+    mkdir build
+    cd build
+    cmake .. -GNinja -DCMAKE_INSTALL_PREFIX=/tmp/gtest_install
+    sudo ninja install
+    rm -rf *
+    cmake .. -GNinja -DCMAKE_CXX_COMPILER=clang++-18 -DCMAKE_CXX_FLAGS=-stdlib=libc++ -DCMAKE_INSTALL_PREFIX=/tmp/gtest_install_libcxx
+    sudo ninja install
+    cd ../..
+fi
